dynamic new icon, module entry points, layout changes

+++ app/src/main/java/com/samsung/android/visualeffect/lock/blind/Blind.java
@@ -16,7 +16,7 @@ public class Blind extends ImageView {
         super(context);
         this.blindX = blindX;
         this.blindWidth = blindWidth;
-        this.midPoint = (blindWidth / 2) + blindX;
+        this.midPoint = ((float) blindWidth / 2) + blindX;
         setImageBitmap(pieceBitmap);
     }
