+++ app/src/main/java/com/samsung/android/visualeffect/lock/rippleink/RippleInkEffect.java
@@ -101,7 +101,7 @@ public class RippleInkEffect extends GLTextureView implements IEffectView {
         if (cmd == 0) {
             this.mRenderer.changeBackground((Bitmap) params.get("Bitmap"));
         } else if (cmd == 1) {
-            this.mRenderer.showUnlockAffordance(((Long) params.get("StartDelay")).longValue(), (Rect) params.get("Rect"));
+            this.mRenderer.showUnlockAffordance((Long) params.get("StartDelay"), (Rect) params.get("Rect"));
         }
     }
